File delete function
-Added script for deleting a file
-Added modal box for confirmation when deleting a file
-Added linked JS and php to file delete modal box
-Added check when uploading a file if file with the same already exists -> will now return an alert instead of replacing the file

+++ php/fileUploader.php
@@ -11,7 +11,7 @@ if(($_SERVER["REQUEST_METHOD"] == "POST")) {
 
   if((!isset($_SESSION["GroupID"])) || (!isset($_SESSION["UserID"])) || (!isset($_SESSION["CourseID"]))) {
     $_SESSION["errormsg"] = "Er ging iets fout!";
-    header("Location: redirect.php?home=1");
+    echo "700";
     exit;
   }
   $userID = $_SESSION["UserID"];
@@ -19,9 +19,15 @@ if(($_SERVER["REQUEST_METHOD"] == "POST")) {
   $courseID = $_SESSION["CourseID"];
 
 
+  if(file_exists(("../files/$groupID/$courseID/" . $_FILES['file']['name'])) == false) {
+    move_uploaded_file($_FILES['file']['tmp_name'],
+      "../files/$groupID/$courseID/" . $_FILES['file']['name']);
+  }else{
+    $_SESSION["errormsg"] = "Er ging iets fout!";
+    echo "705";
+    exit;
+  }
 
-  move_uploaded_file($_FILES['file']['tmp_name'],
-    "../files/$groupID/$courseID/" . $_FILES['file']['name']);
 
   echo "$courseID";
 }
